refactor(core): localize plugin readiness
